functionality of select done, working on styling/D3
