+++ src/actions/introspectionActions.js
@@ -1,11 +1,12 @@
 // import each action type into the actions
-import * as actionTypes from '../constants/actionTypes';
+import * as actionTypes from '../constants/actionTypes.js';
 
 // changing schema action, linked by CHANGE_SCHEMA type
-export const changeSchema = (introspection) => {
+export const changeSchema = (introspection, text) => {
   return {
     type: actionTypes.CHANGE_SCHEMA,
     payload: introspection,
+    text: text,
   };
 }
