+++ src/components/Type.jsx
@@ -1,13 +1,19 @@
 import React from 'react';
 
 const Type = (props) => {
-  const { name, fields } = props;
-  const fieldList = fields.map(field => (
+  const { name, fields, possibleTypes } = props;
+  if (fields !== null) {
+    const fieldList = fields.map(field => (
     <tr>
       <td>{field.name}</td>
       <td>{ (field.type.kind === 'SCALAR') ? field.type.kind : (field.type.kind === 'LIST') ? `[${field.type.ofType.name}]` : field.type.name }</td>
     </tr>
-  ))
+  )) 
+  return fieldList
+}
+  // if (!fields) {
+  //   const possibleTypesList = possibleTypes.map(possibleTypes)
+  // }
 
   return (
     <div className="nodeDisplay">
