+++ src/containers/MainContainer.jsx
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
-import ModalComponent from '../components/ModalComponent.jsx';
-import GraphContainer from './graphContainer.jsx';
+import ModalContainer from './ModalContainer.jsx';
+import GraphContainer from './GraphContainer.jsx';
 
 class MainContainer extends Component {
   render() {
     return (
       <div>
         Main Container is made here
-        <ModalComponent />
+        <ModalContainer />
         <GraphContainer />
       </div>
     );
