+++ src/containers/graphContainer.jsx
@@ -4,9 +4,8 @@ import * as errorAction from '../actions/errorActions';
 import * as introspectionAction from '../actions/introspectionActions';
 import * as svgAction from '../actions/svgActions';
 import * as viewAction from '../actions/viewActions';
-
-import panelContainer from './panelContainer.jsx';
-import modalContainer from './modalContainer.jsx';
+import PanelContainer from './PanelContainer.jsx';
+import ModalContainer from './ModalContainer.jsx';
 import Type from '../components/Type.jsx';
 import { FormGroup } from '@material-ui/core';
 
@@ -25,12 +24,23 @@ class GraphContainer extends Component {
 
   render() {
   // filter introspection query result to related type tables
+    // const typeList = this.props.schema.data.__schema.types.filter((type) => {
+    //   return (type.name.charAt(0) !== "_" && type.name.charAt(1) !== "_" && type.kind !== "SCALAR" && type.kind !== "ENUM" && type.name.toLowerCase() !== "mutation")
+    // });
+
     const typeList = this.props.schema.data.__schema.types.filter((type) => {
-      return (type.name.charAt(0) !== "_" && type.name.charAt(1) !== "_" && type.kind !== "SCALAR" && type.kind !== "ENUM" && type.name.toLowerCase() !== "mutation")
+      return (
+        type.name.charAt(0) !== "_" && type.name.charAt(1) !== "_" && 
+        type.kind !== "INPUT_OBJECT" &&
+        type.kind !== "SCALAR" && 
+        type.kind !== "ENUM" && 
+        (type.fields !== null || type.possibleTypes !== null) &&
+        type.name.toLowerCase() !== "mutation")
     });
 
     const nodes = typeList.map(type => {
-      return <Type fields={type.fields} name={type.name} />
+      console.log(type)
+      return <Type fields={type.fields} name={type.name} possibleTypes={type.possibleTypes}/>
     })
 
     return(
