+++ src/styles/App.css
@@ -1,9 +1,10 @@
-body {
-  background-color: #2F2F2F;
-}
 
 h1 {
   color: #27aedb;
   text-align: center;
 }
 
+body {
+  background-color: #2F2F2F;
+  color: white;
+}
