+++ webpack.config.js
@@ -13,7 +13,14 @@ module.exports = {
       {
         test: /.(js|jsx)$/,
         exclude: /node_modules/,
-        use: ["babel-loader"]
+        use: [
+          {
+            loader: 'babel-loader',
+            options: {
+              compact: false,
+            },
+          },
+        ],
       },
       {
         test: /\.css$/,
